Increase sysinfo ports inode size

+++ src/data_provider/src/ports/iportWrapper.h
@@ -42,7 +42,7 @@ class IPortWrapper
         virtual int32_t remotePort() const = 0;
         virtual int32_t txQueue() const = 0;
         virtual int32_t rxQueue() const = 0;
-        virtual int32_t inode() const = 0;
+        virtual int64_t inode() const = 0;
         virtual std::string state() const = 0;
         virtual int32_t pid() const = 0;
         virtual std::string processName() const = 0;
